labs/lab_25_monitoring/requirements.txt
pytest==7.4.3
rich==13.7.0
psutil==5.9.8
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
apscheduler==3.10.4
slack-sdk==3.27.1
twilio==8.10.0
smtplib3==0.0.1 